End of the section 5. Added identity server.

+++ AuctionService/Controllers/AuctionsController.cs
@@ -1,11 +1,11 @@
-﻿using System.Collections;
-using AuctionService.Data;
+﻿using AuctionService.Data;
 using AuctionService.DTOs;
 using AuctionService.Entities;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -59,11 +59,20 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        auction.Seller = "test";
+        
+        if (User.Identity is not null && User.Identity.Name is not null)
+        {
+            auction.Seller = User.Identity.Name;
+        }
+        else
+        {
+            return Forbid();
+        }
         await _context.Auctions.AddAsync(auction);
         
         var outputAuction = _mapper.Map<AuctionDto>(auction);
@@ -84,6 +93,8 @@ public class AuctionsController : ControllerBase
         return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, outputAuction);
     }
 
+    
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction([FromRoute] Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -94,6 +105,10 @@ public class AuctionsController : ControllerBase
 
         if (auction is null)
             return NotFound();
+
+
+        if (auction.Seller != User.Identity?.Name)
+            return Forbid();
         
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
@@ -116,6 +131,7 @@ public class AuctionsController : ControllerBase
         return BadRequest("Could not save changes.");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction([FromRoute] Guid id)
     {
@@ -123,6 +139,11 @@ public class AuctionsController : ControllerBase
         if (auction is null)
             return Ok();
 
+        if (auction.Seller != User.Identity?.Name)
+        {
+            return Forbid();
+        }
+        
         _context.Auctions.Remove(auction);
 
         var auctionDeleted = new AuctionDeleted()
